don't unassign on a SIGTERM, it's ok to restart and resume in that case
